refactor: Group router-related functions in a folder
